Add dashboard favicon and unify release upload behind New release modal.
Replace separate Upload release button and inline CLI builder with a wizard
modal (Via CLI or Bundle upload), add the star logomark favicon, and extract
shared upload form logic for reuse.
