+++ web-dashboard/index.html
@@ -8,6 +8,7 @@
       content="Codemagic Patch dashboard for managing over-the-air React Native updates."
     />
     <title>Codemagic Patch Dashboard</title>
+    <link rel="icon" href="/favicon.svg" type="image/svg+xml" />
   </head>
   <body>
     <div id="root"></div>
